🐛 Fix: 무한 루프 수정
